Splinting big ball of mud into more manageable chunks.
